# Apply matchers when fetching label values