--- a/tsdb/querier.go
+++ b/tsdb/querier.go
@@ -338,6 +338,24 @@ func labelValuesWithMatchers(r IndexReader, name string, matchers ...*labels.Mat
 	if err != nil {
 		return nil, errors.Wrapf(err, "fetching values of label %s", name)
 	}
+
+	// If we have a matcher for the label name, we can filter out values that don't match
+	// before we fetch postings. This is especially useful for labels with many values.
+	// e.g. __name__ with a selector like {__name__="xyz"}
+	for _, m := range matchers {
+		if m.Name != name {
+			continue
+		}
+
+		for i := len(allValues) - 1; i >= 0; i-- {
+			if m.Matches(allValues[i]) {
+				continue
+			}
+
+			allValues = append(allValues[:i], allValues[i+1:]...)
+		}
+	}
+
 	valuesPostings := make([]index.Postings, len(allValues))
 	for i, value := range allValues {
 		valuesPostings[i], err = r.Postings(name, value)